[CDF-22374] 🦈Move Feature Flag into cdf.toml (#872)
* refactor: added feature flags to cdf.toml

* build: fix

* refactor: added feature flags

* refactor: control feature flag from cdf.toml

* refactor: moved feature flags to its own module

* fix; introduced bug

* build: pytest config
