+++ cdf.toml
@@ -1,6 +1,13 @@
 [cdf]
 project_dir = "cognite_toolkit"
 
+[cdf.feature_flags]
+graphql = true
+asset-dump = true
+timeseries-dump = true
+import-cmd = true
+modules-app = true
+
 [modules]
 # This is the version of the modules. It should not be changed manually.
 # It will be updated by the 'cdf module upgrade' command.
